Separate data from seed and apply runInSeq to reviews
